simple-cart-service/requirements.txt
grpcio==1.62.1
grpcio-tools==1.62.1
redis==5.0.1
